add logging and cleaned parser
